feat: IFS-2400 rename default implementations with IsyOAuth2* prefix
* indicate class is isyfact specific and OAuth2 related
